* rename training script
