+++ training/distributed.py
@@ -1,9 +1,7 @@
-import math
 import pickle
 
 import torch
 from torch import distributed as dist
-from torch.utils.data.sampler import Sampler
 
 
 def get_rank():
